+++ changelog.md
@@ -7,6 +7,7 @@
 * [1387](https://github.com/zeta-chain/node/pull/1387) - Add HSM capability for zetaclient hot key
 * enable zetaclients to use dynamic gas price on zetachain - enables >0 min_gas_price in feemarket module
 * add static chain data for Sepolia testnet
+* added metrics to track the burn rate of the hotkey in the telemetry server as well as prometheus
 
 ### Fixes
 * fix go-staticcheck warnings for zetaclient